Corrected German language issues
Correction of grammatical and spelling mistakes.
Also changed "lux" into "lx" in the en/nl/sv/de.json files.
Like °C or %, is lx the SI unit sign. "Lux" is the unit name, comparable to "temperature" or "humidity". I hope this was ok for you.

And I hope it works now. If not, please let me know.